sanitize pasted content from ms office
