+++ CHANGELOG-1.x.md
@@ -3,6 +3,7 @@
 
 ## v1.0.1
 * When creating a new teaser that has a target, default sync_with_target to true.
+* Blocksmith :: Sanitize pasted content so copy pasta from MS Office products doesn't result in extra empty paragraphs.
 
 
 ## v1.0.0
