+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@triniti/cms",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "license": "Apache-2.0",
   "description": "A single page app for managing triniti schemas and services.",
   "repository": {
