+++ src/blocksmith/plugins/BlocksmithPlugin.js
@@ -6,7 +6,8 @@ import {
   $getSelection,
   $getRoot,
   createCommand,
-  COMMAND_PRIORITY_EDITOR
+  COMMAND_PRIORITY_EDITOR,
+  PASTE_COMMAND,
 } from 'lexical';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import { $getNearestBlockElementAncestorOrThrow, mergeRegister } from '@lexical/utils';
@@ -19,12 +20,14 @@ import blocksToEditor from '@triniti/cms/blocksmith/utils/blocksToEditor.js';
 import editorToBlocks from '@triniti/cms/blocksmith/utils/editorToBlocks.js';
 import getSelectedNode from '@triniti/cms/blocksmith/utils/getSelectedNode.js';
 import marshalToFinalForm from '@triniti/cms/blocksmith/utils/marshalToFinalForm.js';
+import sanitizeNodes from '@triniti/cms/blocksmith/utils/sanitizeNodes.js';
 
 export const INSERT_BLOCK_COMMAND = createCommand();
 export const REMOVE_BLOCK_COMMAND = createCommand();
 export const REPLACE_BLOCK_COMMAND = createCommand();
 export const BLOCKSMITH_DIRTY = 'blocksmith.dirty';
 export const BLOCKSMITH_HYDRATION = 'blocksmith.hydration';
+export const BLOCKSMITH_SANITIZE = 'blocksmith.sanitize';
 
 export default function BlocksmithPlugin(props) {
   const { name = 'blocks', isVisible = false } = props;
@@ -132,6 +135,12 @@ export default function BlocksmithPlugin(props) {
           $node.replace($createBlocksmithNode(curie, newPbj.toObject()));
         }
         return true;
+      }, COMMAND_PRIORITY_EDITOR),
+      editor.registerCommand(PASTE_COMMAND, () => {
+        setTimeout(() => {
+          sanitizeNodes(editor);
+        });
+        return false;
       }, COMMAND_PRIORITY_EDITOR),
     );
   }, [editor]);
