+++ src/plugins/raven/RavenServer.js
@@ -2,6 +2,7 @@ import base64 from 'base-64';
 import utf8 from 'utf8';
 import { serviceIds } from '@triniti/cms/plugins/raven/constants.js';
 import getAccessToken from '@triniti/cms/plugins/iam/selectors/getAccessToken.js';
+import isAuthenticated from '@triniti/cms/plugins/iam/selectors/isAuthenticated.js';
 
 const LOG_PREFIX = `raven_server.v${APP_VERSION}/`;
 
@@ -145,6 +146,10 @@ export default class RavenServer {
 
     this.#prevErrorMessage = error.message;
     console.error(`${LOG_PREFIX}onError`, evt, error);
+
+    if (!this.#app.select(isAuthenticated, true)) {
+      return;
+    }
 
     const data = {
       ctx_app: {
